server/requirements.txt
bcrypt>=4.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.0.1
Flask-CORS==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.0.0
redis==4.6.0
Werkzeug==2.0.1
Pillow==10.0.0
python-magic-bin==0.4.14